+++ CHANGELOG.md
@@ -15,7 +15,7 @@ We use _breaking :warning:_ to mark changes that are not backward compatible (re
 ### Added
 
 - [#4299](https://github.com/thanos-io/thanos/pull/4299) Tracing: Add tracing to exemplar APIs.
-- [#4327](https://github.com/thanos-io/thanos/pull/4327) Add env substitution to PathOrContent flag.
+- [#4327](https://github.com/thanos-io/thanos/pull/4327) Add environment variable substitution to all YAML configuration flags.
 
 ### Fixed